Create README.md
With the ongoing Covid-19 pandemic many people have found themselves with anxieties that have created situations in their lives that has ultimately led to depression or some other mental illness that they had never experienced before. The most important guidance since the beginning of the pandemic has been primarily to stay busy/to keep oneself occupied the best they can in order to deter themselves from falling into any type of mental health issue. This is what the primary focus is of this project. That is, contributing and compiling ideas to continually offset the possibility of one suffering from a mental health issue. These ideas range from simple things such as going for a walk two to three times a day to things more complex and mind and time absorbing as creating and starting ones own online business.